+++ egui_glium/src/backend.rs
@@ -89,7 +89,8 @@ pub fn run(
             };
             app.ui(&ctx, &mut integration_context);
             let app_output = integration_context.output;
-            let (egui_output, paint_jobs) = ctx.end_frame();
+            let (egui_output, paint_commands) = ctx.end_frame();
+            let paint_jobs = ctx.tesselate(paint_commands);
 
             let frame_time = (Instant::now() - egui_start).as_secs_f64() as f32;
             previous_frame_time = Some(frame_time);