+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * You can no longer throw windows
 * `Context::begin_frame()` no longer returns anything.
   * Put your widgets into a `SidePanel`, `TopPanel`, `CentralPanel`, `Window` or `Area`.
+* `Context::end_frame()` now returns "paint jobs" that need to be converted to triangles with `Context::tesselate()`.
 
 ## 0.2.0 - 2020-10-10